clean files and fix dashboard project no data bug
